+++ lib/br_invoices_pdf/cfe/parser/document_number.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 module BrInvoicesPdf
   module Cfe
     module Parser
       module DocumentNumber
-        extend BaseParser
+        extend Util::XmlLocate
 
         module_function