+++ lib/br_invoices_pdf/nfce/renderer/taxes_info.rb
@@ -8,16 +8,14 @@ module BrInvoicesPdf
 
         def execute(pdf, data)
           box(pdf, [0, pdf.cursor], page_content_width(pdf)) do
-            tribute_values(pdf, data[:taxes])
+            tribute_values(pdf, data[:additional_info])
           end
         end
 
         # :reek:FeatureEnvy
         def tribute_values(pdf, taxes)
           pdf.text("Tributos\n\n", style: :italic)
-          value = format_currency(taxes[:amount])
-          msg = 'Informação dos tributos totais incidentes (Lei Federal 12.741/2012)'
-          text = "#{msg}:\n R$ #{value} (#{taxes[:percent]})\n\n"
+          text = "#{taxes}\n\n"
           pdf.text(text, align: :center)
         end
         private_class_method :tribute_values