+++ spec/lib/br_invoices_pdf/cfe/parser/payments_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe BrInvoicesPdf::Cfe::Parser::Payments do
   describe '.execute' do
     subject { described_class.execute(xml) }
@@ -10,7 +12,7 @@ describe BrInvoicesPdf::Cfe::Parser::Payments do
     let(:amount) { 99.9 }
 
     def locate_element_mock(path, value)
-      base_parser_module = BrInvoicesPdf::Cfe::Parser::BaseParser
+      base_parser_module = BrInvoicesPdf::Util::XmlLocate
       allow_any_instance_of(base_parser_module).to receive(:locate_element)
         .with(element, path).and_return(value)
     end