+++ spec/lib/br_invoices_pdf/cfe/parser/products_data_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: false
+
 describe BrInvoicesPdf::Cfe::Parser::ProductsData do
   describe '.execute' do
     subject { described_class.execute(xml) }
@@ -15,7 +17,7 @@ describe BrInvoicesPdf::Cfe::Parser::ProductsData do
     let(:locate_response) { [element] }
 
     def locate_element_mock(path, value)
-      base_parser_module = BrInvoicesPdf::Cfe::Parser::BaseParser
+      base_parser_module = BrInvoicesPdf::Util::XmlLocate
       allow_any_instance_of(base_parser_module).to receive(:node_locate)
         .with(element, path).and_return(value)
     end