+++ spec/lib/br_invoices_pdf/cfe/renderer/payment_forms_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe BrInvoicesPdf::Cfe::Renderer::PaymentForms do
   describe '.execute' do
     subject { described_class.execute(pdf, data) }