+++ spec/lib/br_invoices_pdf/version_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe BrInvoicesPdf::VERSION do
   it 'has a version number' do
     is_expected.to_not be_nil